requirements.txt
flask
pillow
uvicorn
fastapi